requirements.txt
Django==4.0.6
django-cors-headers==3.13.0
mysqlclient==2.1.1
PyJWT==2.4.0
requests==2.28.1
